requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
rank-bm25
openai
torch
transformers
scikit-learn